"6 oct"

+++ backend/controllers/post.js
@@ -30,8 +30,7 @@ export const createPost = async (req, res, next) => {
     try {
         const {title , description} = req.body;
         let imagePath = req.protocol + '://' + req.get("host") + "/images/" + req.file.filename ;
-        console.log({title , description , imagePath})
-        const post = await Post.create({title , description , imagePath});
+        const post = await Post.create({title , description , imagePath , creator:req.user.id});
         res.status(201).json(post);
     } catch (error) {
         next(error);
@@ -42,12 +41,12 @@ export const createPost = async (req, res, next) => {
 export const deletePost = async (req, res, next) => {
     try {
         const { id } = req.params;
-        if (!id) throw new Error("Id is mising");
-        const { deletedCount } = await Post.deleteOne({ _id: id })
-        if (!deletedCount) {
-          return  res.status(404).send("post not found")
+        if (!id) throw new Error("Id is missing");
+        const { deletedCount } = await Post.deleteOne({ _id: id, creator:id})
+        if (deletedCount === 0) {
+          return  res.status(401).send("your post not found")
         }
-        res.status(204).send("succesfully deleted")
+        res.status(204).send("succe ssfully deleted")
     } catch (error) {
         next(error);
     }
@@ -61,14 +60,24 @@ export const updatePost = async (req, res, next) => {
        if(!mongoose.Types.ObjectId.isValid(id)){
          return res.status(400).json({ message: "Invalid ID format" });
        }
-
-        const updatedPost = await Post.findByIdAndUpdate(id,req.body,{new: true});
-
-        if(!updatedPost){
-            return  res.status(404).send("post not found")
+       
+       const {title , description} = req.body;
+       const updatedPostObj = {
+            title , description
+       }
+       if(req?.file?.filename){
+        updatedPostObj.imagePath = req.protocol + '://' + req.get("host") + "/images/" + req.file.filename
+       }else{
+        updatedPostObj.imagePath = req.body.imagePath
+       }
+       
+        const {modifiedCount} = await Post.updateOne({_id:id , creator:id},updatedPostObj);
+        
+        if(modifiedCount === 0){
+            return  res.status(401).send("post not found")
         }
 
-        res.status(200).json("sucessfully updated post");
+        res.status(200).json("successfully updated post");
 
     } catch (error) {
         next(error);
